move action controller that aren't process_action into separate file
